Nnnn range data converter
added a new data converter by request from Duo.
It defines an input range and maps it to an output range.
It includes some extra features like clamping the value, applying a modulo in order to make it loop, and reversing it.
And it also takes an interpolator such that the mapping doesn't need to be lineal.
@

https://github.com/user-attachments/assets/d18b1269-8b5e-4aad-9504-9f5f3dcdaad9

Diffs=
88543fa792 Nnnn range data converter (#8585)