Update linear animation advanceAndApply to return if more animation remains
Origin: using the iOS runtime, playing a linear animation (i.e an animation without a state machine) would not continue playing after the first frame, because `advanceAndApply` was returning `false`, due to the delta initially being 0 and hitting a certain codepath. After a live chat, Luigi and Hernan discussed a usable approach, which resulted in this pull request.

Diffs=
81709286ea Update linear animation advanceAndApply to return if more animation remains (#8610)