GL fixes for Imagination PowerVR GPUs
Don't selectively declare the clipBuffer based on whether ENABLE_CLIPPING is defined. Within a single render pass, some draws have clipping and some don't. Defining the pixel local store differently for different draws within the same render pass corrupted it.

Add guards to ensure a * 1/a == 1 when unmultiplying the dstColor for advanced blend. The branching in the blend equations relies on this being the case.

Diffs=
302606642c GL fixes for Imagination PowerVR GPUs (#8684)