editor+runtime: fix text hit-testing on a locally transformed text
It turns out you can have a transform that's not worldTransform to affect where the text run is rendered. E.g. You can have
1. a non-zero origin set on the Text
2. the overflow mode is 'fit', which first scales the text, then world transforms it.

This PR accounts for this in hit testing. I made sure that when there's a non-zero origin set on a text, a hover on the text run works. Also, if the overflow mode is 'fit', things also work.

Known behavior that's undesirable. If the overflow mode is fit or ellipsis, or basically something that affects whether the text run is rendered at all or changes the bounding box of the run considerably without overlap, then hovering over the edges of the text run could result in flicker (due to hit test toggling constantly). I'm not sure of the best fix for this, but maybe users don't hit this very often either.

Diffs=
7c018ff174 editor+runtime: fix text hit-testing on a locally transformed text (#8860)