fix bad property in animation at load
When an animation tries to key a property that is not supported on the object, the runtime fails the load op. This seemed like a good idea at the time but it affects compatibility as we could remove keyed properties or have new ones that old runtimes do not understand. The runtime should just prune these from the keyed list, which is what this PR does and adds a test with a file that tries to key the old feather value on the shape paint (which we've now removed in favor of one on the feather object itself).

https://2dimensions.slack.com/archives/C07VD44ASE5/p1737664856604919

Diffs=
4c1da5310e fix bad property in animation at load (#8927)