Improve layout display change handling
This fixes reports of a flicker occuring when keying layout display between show/hide. This only occurs when using a nested artboard in the editor and seems to be due to LayoutComponent.isDisplayHidden being queried after changing the display value via LayoutComponent.syncStyle but before updateLayoutBounds is called. When this happens, we were previously just returning the style's display value which could be out of sync if updateLayoutBounds hadn't been called yet.

This may require additional investigation into the layout component calls between dart and CPP since it only occurs via rive native (ie, not when using layouts in an artboard with no nested artboards and also not when viewing at runtime with nested artboards).

Diffs=
0b17404745 Improve layout display change handling (#8951)