Adapt ScrollConstraint to work with NestedArtboardLayouts
This PR adapts ScrollConstraint to work with the generic LayoutNodeProvider. Previously, scrolling only worked with LayoutComponents as children. This change allows NestedArtboardLayouts to be affected by ScrollConstraints, so they can scroll in the same way that LayoutComponents do today. This also lays the groundwork for component lists to be scrollable.

Diffs=
71b138cf40 Adapt ScrollConstraint to work with NestedArtboardLayouts (#8968)