Make the segmentation of feathered curves more uniform
The old method would have tiny pops and wobbles when the segmentation snapped while animating a feathered curve. With the recursion it also wasn't well suited to being moved to the GPU.

Redo the segmentation based entirely on uniform steps in tangent angle. As feather grows initially, we need smaller segments, but once we cross a certain threshold, we can start making the segments bigger again.

This introduces precision issues in some places where more joins are overlapping, but in these scenarios our end goal will be to render to a scaled-down, offscreen atlas, and this atlas can be fp32.

The uniform steps in tangent angle will also be very easy to move to the GPU.

Diffs=
7968fd06b8 Make the segmentation of feathered curves more uniform (#9008)